Put validateLogin into its on utils folder
